+++ src/Contracts/CepRepositoryContract.php
@@ -2,6 +2,8 @@
 
 namespace Atiladanvi\CepRepository\Contracts;
 
+use Atiladanvi\CepRepository\Address;
+
 /**
  * Interface CepRepositoryContract
  * @package Atiladanvi\CepRepository\Contracts
@@ -10,7 +12,7 @@ interface CepRepositoryContract
 {
     /**
      * @param string $cep
-     * @return array
+     * @return ?Address
      */
-    public function get(string $cep);
+    public function get(string $cep): ?Address;
 }