+++ src/Repositories/CepWEBMANIARepository.php
@@ -3,12 +3,22 @@
 namespace Atiladanvi\CepRepository\Repositories;
 
 use Atiladanvi\CepRepository\Clients\WEBMANIAClient;
-use Atiladanvi\CepRepository\Fractals\WEBMANIAFractal;
 
 class CepWEBMANIARepository extends CepRepositoryAbstract
 {
     public function __construct()
     {
-        parent::__construct(WEBMANIAClient::class, WEBMANIAFractal::class);
+        parent::__construct(WEBMANIAClient::class);
+    }
+
+    public function transform($data): array
+    {
+        return [
+            'cep' => $data->cep,
+            'estado' => $data->uf,
+            'municipio' => $data->cidade,
+            'bairro' => $data->bairro,
+            'logradouro' => $data->endereco,
+        ];
     }
 }