+++ tests/CEPLARepositoryTest.php
@@ -10,6 +10,6 @@ class CEPLARepositoryTest extends TestCase
     {
         $ceplaRepository =  app(CepPOSTMONRepository::class);
         $address = $ceplaRepository->get(66911030);
-        $this->assertTrue($address['bairro'] === 'Maracajá (Mosqueiro)');
+        $this->assertTrue($address->bairro === 'Maracajá (Mosqueiro)');
     }
 }