+++ CHANGELOG.md
@@ -6,10 +6,21 @@ All notable changes to `cep-repository` will be documented in this file
 
 - Add initial service architecture
 
-## 0.0.2 - 2021-05-29
+## 0.0.2 - 2020-05-29
 
 - Add support to laravel 6
 
-## 0.0.3 - 2021-06-03
+## 0.0.3 - 2020-06-03
 
 - Add support to laravel 7
+
+## 0.1.0 - 2021-13-04
+
+- Breaking changes:
+  * The get method now return the type:
+    `Atiladanvi\CepRepository\Address`
+    that add more power to handle news features
+- Features:
+   * Address object class added
+- Enhancements:
+   * Degranularize transforms/fractal